Refactor reducer function to separate file
