+++ src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useEffect, useCallback, useReducer } from 'react';
 import axios from 'axios';
-
 import './App.scss';
+import apiReducer from './state';
 
 // Let's talk about using index.js and some other name in the component folder.
 // There's pros and cons for each way of doing this...
@@ -22,33 +22,6 @@ const apiInitialState = {
   loading: false,
   request: {},
   history: []
-};
-
-const apiReducer = (state, action) => {
-  switch (action.type) {
-    case 'SET_REQUEST_PARAMETERS':
-      return { ...state, requestParams: action.payload };
-    case 'SET_LOADING':
-      return { ...state, loading: action.payload };
-    case 'SET_DATA':
-      return { ...state, data: action.payload };
-    case 'SET_REQUEST':
-      return { ...state, request: action.payload };
-    case 'ADD_TO_HISTORY': {
-      const exists = state.history.some(
-        (item) =>
-          item.method === action.payload.method &&
-          item.url === action.payload.url &&
-          JSON.stringify(item.data) === JSON.stringify(action.payload.data)
-      );
-      if (!exists) {
-        return { ...state, history: [...state.history, action.payload] };
-      }
-      return state;
-    }
-    default:
-      return state;
-  }
 };
 
 const App = () => {
